backend/requirements.txt
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.13.0
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
distro==1.9.0
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.17.1
docstring_parser==0.17.0
exceptiongroup==1.3.1
gunicorn==25.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
packaging==26.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
requests==2.33.0
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
